Changes FPHA model to PWL-4
First experimental change to FPHA modelling. Let's start by making it a piecewiese linear aproximation of the FPH by a model that don't use binarie variables, suitable for Linear Problems. We are still using HiGhs as the solver in this version. For future versions we expect to turn the problem into a MILP and change the solver to Gurobi. For that, we are gonna need a new model for the FPHA. Summarizing, this patch is for adding a experimental model that will be changed in the future.